+++ examples/simulation-study-aggregated-observations/src/plot.R
@@ -4,119 +4,111 @@ library(magrittr)
 library(ggplot2)
 library(stringr)
 library(jsonlite)
+library(coda)
 
 green_hex_colour <- "#7fc97f"
 purple_hex_colour <- "#beaed4"
 
-## =============================================================================
-## Generate cross sections of the LLHD function in the birth rate.
-## =============================================================================
-x <- list(
-  "out/llhd-evaluations-est-params-agg-data.csv",
-  "out/llhd-evaluations-est-params-regular-data.csv",
-  "out/llhd-evaluations-true-params-regular-data.csv"
-)
+
+SAVE_FIGURES <- TRUE
 
 app_config <- read_json("agg-app-config.json")
 sim_lambda <- app_config$simulationParameters[[1]]
+sim_duration <- app_config$simulationDuration
 
-read_llhds <- function(filename) {
-  if (file.exists(filename)) {
-    filename %>%
-      read.csv(header = FALSE) %>%
-      set_names(c("type", "value")) %>%
-      mutate(lambda = seq(from = 1, to = 3.5, length = 100))
-  } else {
-    NULL
-  }
-}
 
-cross_section_df <- map(
-  x,
-  read_llhds
-) %>%
-  keep(compose(not, is_null)) %>%
-  bind_rows() %>%
+## =============================================================================
+## Generate a figure looking at the posterior samples conditioned upon the
+## regular data, i.e., the unscheduled observations.
+## =============================================================================
+
+reg_data_mcmc_csv <- app_config$inferenceConfigurations %>%
+  extract2(2) %>%
+  extract("icMaybeMCMCConfig") %>%
+  extract2(1) %>%
+  extract2("mcmcOutputCSV")
+
+reg_data_mcmc_df <- read.csv(reg_data_mcmc_csv) %>%
   mutate(
-    estimated_params = str_detect(type, "estimated.*"),
-    aggregated_data = str_detect(type, ".*aggregated.*")
+    nb_min = qnbinom(p = 0.025, size = nbSize, prob = 1 - nbProb),
+    nb_med = qnbinom(p = 0.5, size = nbSize, prob = 1 - nbProb),
+    nb_max = qnbinom(p = 0.975, size = nbSize, prob = 1 - nbProb)
   )
 
-g <- ggplot(cross_section_df) +
-  geom_line(mapping = aes(x = lambda, y = value, colour = aggregated_data, linetype = estimated_params)) +
-  geom_vline(xintercept = sim_lambda) +
-  facet_wrap(~type, scales = "free_y") +
-  scale_linetype_manual(values = c("dashed", "solid")) +
-  scale_color_manual(values = c(green_hex_colour, purple_hex_colour)) +
-  labs(y = NULL, x = "Birth rate (lambda)") +
-  theme_classic() +
-  theme(axis.title = element_text(face = "bold"),
-        legend.position = "none")
+reg_data_nb_summary <- reg_data_mcmc_df %>%
+  select(starts_with("nb_")) %>%
+  colMeans() %>%
+  as.list() %>%
+  as.data.frame()
+reg_data_nb_summary$absolute_time <- sim_duration
 
-ggsave("out/lambda-llhd-cross-sections.png", g)
-
-fig_height <- 8
-ggsave("out/lambda-llhd-cross-sections.png",
-       g,
-       height = fig_height,
-       width = 2 * 1.618 * fig_height,
-       units = "cm"
-       )
-ggsave("out/lambda-llhd-cross-sections.pdf",
-       g,
-       height = fig_height,
-       width = 2 * 1.618 * fig_height,
-       units = "cm"
-       )
-
-## =============================================================================
-## Generate a figure looking at the posterior distribution of the prevalence
-## =============================================================================
-data_paths <- list(
-  "out/final-negative-binomial-est-params-agg-data.csv",
-  "out/final-negative-binomial-est-params-regular-data.csv",
-  "out/final-negative-binomial-true-params-regular-data.csv"
-)
-
-read_nb <- function(filename) {
-  if (file.exists(filename)) {
-    foo <- filename %>%
-      readLines() %>%
-      str_split(pattern = "(,| )") %>%
-      unlist()
-
-    percentile_probs <- c(0.025, 0.25, 0.5, 0.75, 0.975)
-    percentile_vals <- qnbinom(
-      p = c(0.025, 0.25, 0.5, 0.75, 0.975),
-      size = as.numeric(foo[3]),
-      prob = 1 - as.numeric(foo[4])
-    )
-    estimate_type <- foo[1]
-
-    data.frame(
-      percentile_prob = percentile_probs,
-      percentile_value = percentile_vals,
-      estimate_name = estimate_type
-    )
-  } else {
-    NULL
-  }
+small_mcmc_subset <- if (nrow(reg_data_mcmc_df) > 1000) {
+  sample_n(reg_data_mcmc_df, 1000)
+} else {
+  reg_data_mcmc_df
 }
 
-posterior_plot_df <- map(data_paths, read_nb) %>%
-  keep(compose(not, is_null)) %>%
-  bind_rows()
+if (SAVE_FIGURES) {
+  png("out/regular-data-mcmc-pairs-plot.png")
+  pairs(select(small_mcmc_subset, llhd, lambda, psi, omega))
+  dev.off()
+}
 
-g <- ggplot(posterior_plot_df) +
-  geom_point(mapping = aes(
-    x = estimate_name,
-    y = percentile_value,
-    colour = percentile_prob
-  )) +
-  ylim(c(0, 1.1 * max(posterior_plot_df$percentile_value)))
+reg_data_mcmc <- mcmc(select(reg_data_mcmc_df, llhd, lambda, psi, omega))
+
+if (SAVE_FIGURES) {
+  png("out/regular-data-mcmc-trace.png")
+  plot(reg_data_mcmc)
+  dev.off()
+}
+
+## =============================================================================
+## Generate a figure looking at the posterior samples conditioned upon the
+## aggregated data, i.e., the observations that have been aggregated into
+## scheduled observations.
+## =============================================================================
+
+agg_data_mcmc_csv <- app_config$inferenceConfigurations %>%
+  extract2(3) %>%
+  extract("icMaybeMCMCConfig") %>%
+  extract2(1) %>%
+  extract2("mcmcOutputCSV")
+
+agg_data_mcmc_df <- read.csv(agg_data_mcmc_csv) %>%
+  mutate(
+    nb_min = qnbinom(p = 0.025, size = nbSize, prob = 1 - nbProb),
+    nb_med = qnbinom(p = 0.5, size = nbSize, prob = 1 - nbProb),
+    nb_max = qnbinom(p = 0.975, size = nbSize, prob = 1 - nbProb)
+  )
 
 
-ggsave("out/posterior-prevelance-estimates.png", g)
+agg_data_nb_summary <- agg_data_mcmc_df %>%
+  select(starts_with("nb_")) %>%
+  colMeans() %>%
+  as.list() %>%
+  as.data.frame()
+agg_data_nb_summary$absolute_time <- sim_duration
+
+small_mcmc_subset <- if (nrow(agg_data_mcmc_df) > 1000) {
+  sample_n(agg_data_mcmc_df, 1000)
+} else {
+  agg_data_mcmc_df
+}
+
+
+if (SAVE_FIGURES) {
+  png("out/aggregated-data-mcmc-pairs-plot.png")
+  pairs(select(small_mcmc_subset, llhd, lambda, rho, nu))
+  dev.off()
+}
+
+agg_data_mcmc <- mcmc(select(agg_data_mcmc_df, llhd, lambda, rho, nu))
+
+if (SAVE_FIGURES) {
+  png("out/aggregated-data-mcmc-trace.png")
+  plot(agg_data_mcmc)
+  dev.off()
+}
 
 ## =============================================================================
 ## Generate a figure looking at the prevalence through time and the data used in
@@ -210,20 +202,6 @@ agg_occ_df <- aggregated_data %>%
 
 ## -----------------------------------------------------------------------------
 
-sim_duration <- app_config$simulationDuration
-
-tmp_true_regular <- filter(posterior_plot_df, estimate_name == "true_parameters_regular_data") %>% use_series("percentile_value")
-true_regular_df <- data.frame(absolute_time = sim_duration, prev_est_min = min(tmp_true_regular), prev_est_mid = median(tmp_true_regular), prev_est_max = max(tmp_true_regular))
-
-tmp_est_regular <- filter(posterior_plot_df, estimate_name == "estimated_parameters_regular_data") %>% use_series("percentile_value")
-est_regular_df <- data.frame(absolute_time = sim_duration, prev_est_min = min(tmp_est_regular), prev_est_mid = median(tmp_est_regular), prev_est_max = max(tmp_est_regular))
-
-tmp_est_aggregated <- filter(posterior_plot_df, estimate_name == "estimated_parameters_aggregated_data") %>% use_series("percentile_value")
-est_aggregated_df <- data.frame(absolute_time = sim_duration, prev_est_min = min(tmp_est_aggregated), prev_est_mid = median(tmp_est_aggregated), prev_est_max = max(tmp_est_aggregated))
-
-
-
-
 
 g <- ggplot() +
   geom_step(data = prev_df, mapping = aes(x = absolute_time, y = prevalence)) +
@@ -233,33 +211,23 @@ g <- ggplot() +
   geom_segment(data = agg_occ_df, mapping = aes(x = absolute_time, y = num_obs, xend = absolute_time, yend = 0), colour = purple_hex_colour) +
   geom_point(data = agg_occ_df, mapping = aes(x = absolute_time, y = num_obs), colour = purple_hex_colour) +
   geom_errorbar(
-    data = true_regular_df,
-    mapping = aes(x = absolute_time - 0.1, ymin = prev_est_min, ymax = prev_est_max),
-    colour = green_hex_colour, linetype = "dashed", width = 0.2
-  ) +
-  geom_point(
-    data = true_regular_df,
-    mapping = aes(x = absolute_time - 0.1, y = prev_est_mid),
-    colour = green_hex_colour
-  ) +
-  geom_errorbar(
-    data = est_regular_df,
-    mapping = aes(x = absolute_time, ymin = prev_est_min, ymax = prev_est_max),
+    data = reg_data_nb_summary,
+    mapping = aes(x = absolute_time - 0.1, ymin = nb_min, ymax = nb_max),
     colour = green_hex_colour, linetype = "solid", width = 0.2
   ) +
   geom_point(
-    data = est_regular_df,
-    mapping = aes(x = absolute_time, y = prev_est_mid),
+    data = reg_data_nb_summary,
+    mapping = aes(x = absolute_time - 0.1, y = nb_med),
     colour = green_hex_colour
   ) +
   geom_errorbar(
-    data = est_aggregated_df,
-    mapping = aes(x = absolute_time + 0.1, ymin = prev_est_min, ymax = prev_est_max),
+    data = agg_data_nb_summary,
+    mapping = aes(x = absolute_time + 0.1, ymin = nb_min, ymax = nb_max),
     colour = purple_hex_colour, linetype = "solid", width = 0.2
   ) +
   geom_point(
-    data = est_aggregated_df,
-    mapping = aes(x = absolute_time + 0.1, y = prev_est_mid),
+    data = agg_data_nb_summary,
+    mapping = aes(x = absolute_time + 0.1, y = nb_med),
     colour = purple_hex_colour
   ) +
   labs(y = NULL, x = "Time since origin") +
@@ -267,15 +235,18 @@ g <- ggplot() +
   theme(axis.title = element_text(face = "bold"))
 
 fig_height <- 10
-ggsave("out/regular-and-aggregated-data.png",
-  g,
-  height = fig_height,
-  width = 1.618 * fig_height,
-  units = "cm"
-)
-ggsave("out/regular-and-aggregated-data.pdf",
-  g,
-  height = fig_height,
-  width = 1.618 * fig_height,
-  units = "cm"
-)
+
+if (SAVE_FIGURES) {
+  ggsave("out/regular-and-aggregated-data.png",
+    g,
+    height = fig_height,
+    width = 1.618 * fig_height,
+    units = "cm"
+  )
+  ggsave("out/regular-and-aggregated-data.pdf",
+    g,
+    height = fig_height,
+    width = 1.618 * fig_height,
+    units = "cm"
+  )
+}